Fuel Consumption Application
An application that takes three float variables. Cost, Distance Travelled and Litres used.

The application calculates the Kilometeres per Kilometer. And the cost of the distance travelled based on the cost entered per litre.

Notes:
Can add functionality into the application to switch between KM or Miles. Also can add functionality to change the currency for the cost.